@@ -1,17 +1,20 @@
 class Item {
+  final int productId;
   final String name;
   final String unit;
   final int price;
   final String image;
 
   Item(
-      {required this.name,
+      {required this.productId,
+      required this.name,
       required this.unit,
       required this.price,
       required this.image});
 
   Map toJson() {
     return {
+      'product_id': productId,
       'name': name,
       'unit': unit,
       'price': price,
